Relocate `t` functions to `zshrc-t.zsh`
It's been awful for editing and debugging with one-liner zsh functions.

Their intent was for easy scanning simply functions, which made sense for `g` and `tf` remaps, where there is not much logic.

But for things like `args`, where there is a lot of logic, it's been a bad dev experience.

`t` will be the first fork to start writing properly formatted code for functions with real logic.

One-liner style likely still make sense for `g` and `tf` remaps though.